Linux: Add support for sample-tap packet sampling
sampletap is a packet sampling method that allow to sample packets on a
linux machine and transfer them to the userspace via a tap device. The
sampletap, as opposed to NFLOG and ULOG sampling methods, is created via
the tc tool and can be offloaded by hardware vendors.

sample-tap configuration example can be found under the scripts dir, and
is supported from kernel version 4.10

Signed-off-by: Yotam Gigi <yotamg@mellanox.com>